+++ adapters/generic/motion_sensor.py
@@ -2,7 +2,7 @@ from adapters.adapter_with_battery import AdapterWithBattery
 from devices.sensor.motion import MotionSensor
 
 
-class SensorMotion(AdapterWithBattery):
+class MotionSensorAdapter(AdapterWithBattery):
     def __init__(self, devices):
         super().__init__(devices)
         self.devices.append(MotionSensor(devices, 'motion', 'occupancy'))