+++ adapters/generic/smoke_sensor.py
@@ -2,7 +2,7 @@ from adapters.adapter_with_battery import AdapterWithBattery
 from devices.sensor.smoke import SmokeSensor
 
 
-class SensorSmoke(AdapterWithBattery):
+class SmokeSensorAdapter(AdapterWithBattery):
     def __init__(self, devices):
         super().__init__(devices)
         self.devices.append(SmokeSensor(devices, 'smoke', 'smoke'))